Move status creation into environmentVariableInfo
