+++ src/vs/workbench/contrib/terminal/browser/terminalStatusList.ts
@@ -12,7 +12,7 @@ import { TerminalSettingId } from 'vs/platform/terminal/common/terminal';
 import { listErrorForeground, listWarningForeground } from 'vs/platform/theme/common/colorRegistry';
 import { spinningLoading } from 'vs/platform/theme/common/iconRegistry';
 import { ThemeIcon } from 'vs/base/common/themables';
-import { IHoverAction } from 'vs/workbench/services/hover/browser/hover';
+import { ITerminalStatus } from 'vs/workbench/contrib/terminal/common/terminal';
 
 /**
  * The set of _internal_ terminal statuses, other components building on the terminal should put
@@ -24,29 +24,6 @@ export const enum TerminalStatus {
 	RelaunchNeeded = 'relaunch-needed',
 	EnvironmentVariableInfoChangesActive = 'env-var-info-changes-active',
 	ShellIntegrationAttentionNeeded = 'shell-integration-attention-needed'
-}
-
-export interface ITerminalStatus {
-	/** An internal string ID used to identify the status. */
-	id: string;
-	/**
-	 * The severity of the status, this defines both the color and how likely the status is to be
-	 * the "primary status".
-	 */
-	severity: Severity;
-	/**
-	 * An icon representing the status, if this is not specified it will not show up on the terminal
-	 * tab and will use the generic `info` icon when hovering.
-	 */
-	icon?: ThemeIcon;
-	/**
-	 * What to show for this status in the terminal's hover.
-	 */
-	tooltip?: string | undefined;
-	/**
-	 * Actions to expose on hover.
-	 */
-	hoverActions?: IHoverAction[];
 }
 
 export interface ITerminalStatusList {
